+++ flint/prefect/common/imaging.py
@@ -39,7 +39,7 @@ from flint.ms import (
     rename_column_in_ms,
     split_by_field,
 )
-from flint.naming import FITSMaskNames, processed_ms_format
+from flint.naming import FITSMaskNames, get_beam_resolution_str, processed_ms_format
 from flint.options import FieldOptions
 from flint.peel.potato import potato_peel
 from flint.prefect.common.utils import upload_image_as_artifact
@@ -352,6 +352,7 @@ def task_convolve_image(
     cutoff: float = 60,
     mode: str = "image",
     filter: Optional[str] = None,
+    convol_suffix_str: str = "conv",
 ) -> Collection[Path]:
     """Convolve images to a specified resolution
 
@@ -360,6 +361,7 @@ def task_convolve_image(
         beam_shape (BeamShape): The shape images will be convolved to
         cutoff (float, optional): Maximum major beam axis an image is allowed to have before it will not be convolved. Defaults to 60.
         filter (Optional[str], optional): This string must be contained in the image path for it to be convolved. Defaults to None.
+        convol_suffix_str (str, optional): The suffix added to the convolved images. Defaults to 'conv'.
 
     Returns:
         Collection[Path]: Path to the output images that have been convolved.
@@ -412,7 +414,10 @@ def task_convolve_image(
         )
 
     return convolve_images(
-        image_paths=image_paths, beam_shape=beam_shape, cutoff=cutoff
+        image_paths=image_paths,
+        beam_shape=beam_shape,
+        cutoff=cutoff,
+        convol_suffix=convol_suffix_str,
     )
 
 
@@ -493,13 +498,16 @@ def task_linmos_images(
     return linmos_cmd
 
 
-def _convolve_linmos_residuals(
+def _convolve_linmos(
     wsclean_cmds: Collection[WSCleanCommand],
     beam_shape: BeamShape,
     field_options: FieldOptions,
     linmos_suffix_str: str,
     cutoff: float = 0.05,
     field_summary: Optional[FieldSummary] = None,
+    convol_mode: str = "image",
+    convol_filter: str = "-MFS-",
+    convol_suffix_str: str = "conv",
 ) -> LinmosCommand:
     """An internal function that launches the convolution to a common resolution
     and subsequent linmos of the wsclean residual images.
@@ -511,19 +519,24 @@ def _convolve_linmos_residuals(
         linmos_suffix_str (str): The suffix string passed to the linmos parset name
         cutoff (float, optional): The primary beam attenuation cutoff supplied to linmos when coadding. Defaults to 0.05.
         field_summary (Optional[FieldSummary], optional): The summary of the field, including (importantly) to orientation of the third-axis. Defaults to None.
+        convol_mode (str, optional): The mode passed to the convol task to describe the images to extract. Support image or residual.  Defaults to image.
+        convol_filter (str, optional): A text file applied when assessing images to co-add. Defaults to '-MFS-'.
+        convol_suffix_str (str, optional): The suffix added to the convolved images. Defaults to 'conv'.
 
     Returns:
         LinmosCommand: Resulting linmos command parset
     """
-    residual_conv_images = task_convolve_image.map(
+
+    conv_images = task_convolve_image.map(
         wsclean_cmd=wsclean_cmds,
         beam_shape=unmapped(beam_shape),
         cutoff=150.0,
-        mode="residual",
-        filter="-MFS-",
+        mode=convol_mode,
+        filter=convol_filter,
+        convol_suffix_str=convol_suffix_str,
     )
     parset = task_linmos_images.submit(
-        images=residual_conv_images,
+        images=conv_images,
         container=field_options.yandasoft_container,
         suffix_str=linmos_suffix_str,
         holofile=field_options.holofile,
@@ -532,6 +545,75 @@ def _convolve_linmos_residuals(
     )
 
     return parset
+
+
+def _create_convol_linmos_images(
+    wsclean_cmds: Collection[WSCleanCommand],
+    field_options: FieldOptions,
+    field_summary: Optional[FieldSummary] = None,
+    current_round: Optional[int] = None,
+) -> List[LinmosCommand]:
+    """Derive the approriate set of beam shapes and then produce corresponding
+    convolved and co-added images
+
+    Args:
+        wsclean_cmds (Collection[WSCleanCommand]): Set of wsclean commands that have been executed
+        field_options (FieldOptions): Set of field imaging optins, containing details of the beam/s
+        field_summary (Optional[FieldSummary], optional): Summary of the MSs, importantly containing their third-axis rotation. Defaults to None.
+        current_round (Optional[int], optional): Which self-cal imaging round. If None 'noselfcal'. Defaults to None.
+
+    Returns:
+        List[LinmosCommand]: The collection of linmos commands executed.
+    """
+    parsets: List[LinmosCommand] = []
+    main_linmos_suffix_str = f"round{current_round}" if current_round else "noselfcal"
+
+    todo: List[Any, str] = [(None, get_beam_resolution_str(mode="optimal"))]
+    if field_options.fixed_beam_shape:
+        logger.info(
+            f"Creating second round of linmos images with {field_options.fixed_beam_shape}"
+        )
+        todo.append(
+            (field_options.fixed_beam_shape, get_beam_resolution_str(mode="fixed"))
+        )
+
+    for round_beam_shape, beam_str in todo:
+        linmos_suffix_str = f"{beam_str}.{main_linmos_suffix_str}"
+        convol_suffix_str = f"{beam_str}.conv"
+
+        beam_shape = task_get_common_beam.submit(
+            wsclean_cmds=wsclean_cmds,
+            cutoff=field_options.beam_cutoff,
+            filter="-MFS-",
+            fixed_beam_shape=round_beam_shape,
+        )
+        parset = _convolve_linmos(
+            wsclean_cmds=wsclean_cmds,
+            beam_shape=beam_shape,
+            field_options=field_options,
+            linmos_suffix_str=linmos_suffix_str,
+            cutoff=field_options.pb_cutoff,
+            field_summary=field_summary,
+            convol_mode="image",
+            convol_filter="-MFS-",
+            convol_suffix_str=convol_suffix_str,
+        )
+        parsets.append(parset)
+
+        if field_options.linmos_residuals:
+            _convolve_linmos(
+                wsclean_cmds=wsclean_cmds,
+                beam_shape=beam_shape,
+                field_options=field_options,
+                linmos_suffix_str=f"{linmos_suffix_str}.residual",
+                cutoff=field_options.pb_cutoff,
+                field_summary=field_summary,
+                convol_mode="residual",
+                convol_filter="-MFS-",
+                convol_suffix_str=convol_suffix_str,
+            )
+
+    return parsets
 
 
 @task