+++ flint/validation.py
@@ -499,10 +499,10 @@ def plot_rms_map(
 
     # Pirate no believes get below 10uJy/beam mateeey. Percentile a little more
     # robust to outliers but much of a much
-    floor: float = max(1, np.log10(np.nanpercentile(rms_data * 1e6, 16)))  # type: ignore
+    floor: float = max(1, np.floort(np.log10(np.nanpercentile(rms_data, 16))))  # type: ignore
 
     im = ax.imshow(
-        np.log10(rms_data * 1e6),
+        np.log10(rms_data),
         vmin=floor,
         vmax=floor + 1.0,
         origin="lower",
@@ -530,7 +530,6 @@ def plot_rms_map(
 
     cbar = fig.colorbar(
         im,
-        cax=None,
         ax=ax,
         use_gridspec=True,
         pad=0.09,