+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - copying folders in `copy_files_into` when archiving
 - added reading of ArchiveOptions from strategy file for continuum pipeline and `flint_archive`
 - Adaptive colour bar scaling in the rms validation plot
+- Create multiple linmos images if `--fixed-beam-shape` specified, one at an optimal common resolution and another at the specified resolution
 
 ## 0.2.4